requirements.txt
langchain
langchain-groq
langgraph
python-dotenv
pydantic
ipython
streamlit
pytest
pytest-cov 